pyproject.toml
[project]
name = "multi-agent-system"
version = "1.2.0"
description = "A multi-agent system that handles visa applications"
authors = [
    {name = "Safarveisi",email = "sajad.safarveisi@gmail.com"},
    {name = "bahrami",email = "maryam.bahrami.ai@gmail.com"},
    {name = "Petrosian",email = "celine.petrosian@gmail.com"},
    {name = "Maboody",email = "m.maboody@gmail.com"},
    {name = "Jelodari",email = "iraj.jelo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "haystack-ai (>=2.13.1,<3.0.0)",
    "docstring-parser (>=0.16,<0.17)",
    "streamlit (>=1.45.0,<2.0.0)",
    "jsonschema (>=4.23.0,<5.0.0)",
    "anthropic-haystack (>=2.4.1,<3.0.0)",
    "ollama-haystack (>=2.4.2,<3.0.0)"
]

[tool.poetry]
package-mode = true

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
black = {extras = ["jupyter"], version = "^25.1.0"}
rich = "^14.0.0"
ruff = "^0.11.12"
